refactor: 브랜드 상세 페이지 로직 분리
